+++ database/migrations/2023_03_29_215622_create_author.php
@@ -11,8 +11,8 @@ class CreateAuthor extends Migration
         // 作者信息
         Schema::create('author', function (Blueprint $table) {
             $table->id();
-            $table->integer("uid", false, true)->unique();
-            $table->integer("user_uid", false, true)->unique();
+            $table->bigInteger("uid", false, true)->unique();
+            $table->bigInteger("user_uid", false, true)->unique();
             $table->tinyInteger("is_forbidden", false, true)->default(2);
             $table->integer("series_count", false, true)->default(0);// 作品数量
             $table->string("qr_url", 100)->default("");// 个人二维码
