+++ database/migrations/2023_03_29_220509_create_image.php
@@ -11,11 +11,11 @@ class CreateImage extends Migration
         // 图片基础信息
         Schema::create('image', function (Blueprint $table) {
             $table->id();
-            $table->integer("uid", false, true)->unique();
-            $table->integer("user_uid", false, true);
-            $table->integer("author_uid", false, true);
-            $table->integer("series_uid", false, true);
-            $table->integer("category_uid", false, true);
+            $table->bigInteger("uid", false, true)->unique();
+            $table->bigInteger("user_uid", false, true);
+            $table->bigInteger("author_uid", false, true);
+            $table->bigInteger("series_uid", false, true);
+            $table->bigInteger("category_uid", false, true);
             $table->integer("item_count", false, true)->default(0);// 张数
             $table->string("url", 255);
             $table->string("path", 255);
